Raiden v 0.2
All the code for changing between welcome screen, map 1, map 2 and stage
clear screen has been added. All changes but the final one are done with
spacebar, the change to the stage clear screen has to be done with the
key "d". The welcome screen and stage clear screen have no sound, while
the map 1 and map 2 have each a different song that correspond with the
map that is in the screen (it also changes).