Enemies shoot
Enemies already shoot depending on the proximity of the player. There is
a bug in which the program throws a segmentation fault when you kill the
first three planes. Only when you kill the three of them, otherwise the
program works fine. It seems to be a problem of the particle's collider,
but i don't know how to fix it yet.